requirements.txt
pydub==0.25.1
openai-whisper==20231117